teste_git/cat_facts_collector/v1_cat_fact_official_2026_01/requirements.txt
numpy>=2.1.0
requests>=2.31.0
pandas>=2.2.0
pydantic>=2.5.0
python-dotenv>=1.0.0
colorlog>=6.7.0
urllib3>=2.0.0
